Add test for constraint binding.
